add category "

+++ internal/core/usecase/task_usecase.go
@@ -22,19 +22,19 @@ func NewTaskService(taskRepo *persistent.TaskRepository, userRepo *persistent.Us
 func (s *TaskService) CreateTask(task *entity.Task) error {
 
     if err := s.taskRepo.Create(task); err != nil {
-        return response.NewAppError("500", err.Error())
+        return response.NewAppError(500, err.Error())
     }
     return nil
 }
 
 func (s *TaskService) GetUserTasks(userID uint) ([]entity.Task, error) {
     if !s.userRepo.IsUserExistByID(userID) {
-        return nil, response.NewAppError("404", "User not found")
+        return nil, response.NewAppError(404, "User not found")
     }
 
     tasks, err := s.taskRepo.FindByUserID(userID)
     if err != nil {
-        return nil, response.NewAppError("500", err.Error())
+        return nil, response.NewAppError(500, err.Error())
 
     }
     return tasks, nil
@@ -42,29 +42,51 @@ func (s *TaskService) GetUserTasks(userID uint) ([]entity.Task, error) {
 
 func (s *TaskService) GetSubTasks(taskID uint) ([]entity.Task, error) {
     if !s.taskRepo.IsTaskExistByID(taskID) {
-        return nil, response.NewAppError("404", "Task not found")
+        return nil, response.NewAppError(404, "Task not found")
     }
 
     tasks, err := s.taskRepo.FindSubTasks(taskID)
     if err != nil {
-        return nil, response.NewAppError("500", err.Error())
+        return nil, response.NewAppError(400, err.Error())
     }
 
+    return tasks, nil
+}
+
+func (s *TaskService) GetTaskByCategory(categoryID uint64, userID uint) ([]entity.Task, error) {
+
+    tasks, err := s.taskRepo.FindByCategoryIDAndUserID(categoryID, userID)
+    if err != nil {
+        return nil, response.NewAppError(400, err.Error())
+    }
+    return tasks, nil
+}
+
+func (s *TaskService) GetNonCategoryTasks(userID uint) ([]entity.Task, error) {
+
+    tasks, err := s.taskRepo.FindNonCategoryTask(userID)
+    if err != nil {
+        return nil, response.NewAppError(400, err.Error())
+    }
     return tasks, nil
 }
 
 func (s *TaskService) UpdateTask(task *entity.Task) error {
 
     if err := s.taskRepo.Update(task); err != nil {
-        return response.NewAppError("500", err.Error())
+        return response.NewAppError(500, err.Error())
     }
     return nil
 }
 
 func (s *TaskService) DeleteTask(taskID uint) error {
 
+    if !s.taskRepo.IsTaskExistByID(taskID) {
+        return response.NewAppError(404, "Task not found")
+    }
+
     if err := s.taskRepo.Delete(taskID); err != nil {
-        return response.NewAppError("500", err.Error())
+        return response.NewAppError(500, err.Error())
     }
     return nil
 }
